FIX Forward autocast_adapter_dtype in task-type add_adapter
PeftModelForSequenceClassification, PeftModelForTokenClassification and
PeftModelForQuestionAnswering override add_adapter to inject their head into
modules_to_save. All three accept and document autocast_adapter_dtype but
called super().add_adapter() without passing it on, so PeftModel's default of
True always won: adapters were upcast to float32 even when the user explicitly
passed autocast_adapter_dtype=False.

This also affected PeftModel.load_adapter for these task types, since it
routes through the same override; its trailing cast_adapter_dtype call is a
no-op when the flag is False, so the adapter stayed float32 after loading.

Adds tests covering all three task types, for add_adapter and load_adapter,
asserting both that False keeps the original dtype and that True still upcasts.
This code path previously had no test coverage at all.

+++ src/peft/peft_model.py
@@ -1859,7 +1859,12 @@ class PeftModelForSequenceClassification(PeftModel):
             else:
                 peft_config.modules_to_save.extend(classifier_module_names)
 
-        return super().add_adapter(adapter_name, peft_config, low_cpu_mem_usage=low_cpu_mem_usage)
+        return super().add_adapter(
+            adapter_name,
+            peft_config,
+            low_cpu_mem_usage=low_cpu_mem_usage,
+            autocast_adapter_dtype=autocast_adapter_dtype,
+        )
 
     def forward(
         self,
@@ -2713,7 +2718,12 @@ class PeftModelForTokenClassification(PeftModel):
             else:
                 peft_config.modules_to_save.extend(classifier_module_names)
 
-        return super().add_adapter(adapter_name, peft_config, low_cpu_mem_usage=low_cpu_mem_usage)
+        return super().add_adapter(
+            adapter_name,
+            peft_config,
+            low_cpu_mem_usage=low_cpu_mem_usage,
+            autocast_adapter_dtype=autocast_adapter_dtype,
+        )
 
     def forward(
         self,
@@ -2943,7 +2953,12 @@ class PeftModelForQuestionAnswering(PeftModel):
             else:
                 peft_config.modules_to_save.extend(qa_module_names)
 
-        return super().add_adapter(adapter_name, peft_config, low_cpu_mem_usage=low_cpu_mem_usage)
+        return super().add_adapter(
+            adapter_name,
+            peft_config,
+            low_cpu_mem_usage=low_cpu_mem_usage,
+            autocast_adapter_dtype=autocast_adapter_dtype,
+        )
 
     def forward(
         self,
